aiogram-django/requirements.txt
aiogram~=3.0.0b6
dj_database_url~=1.0.0
Django~=4.1.3
django-split-settings~=1.2.0
django-stubs~=1.13.0
gunicorn~=20.1.0
python-dotenv~=0.21.0
starlette~=0.22.0
uvicorn~=0.19.0
django-import-export~=3.0.1
django-admin-extra-buttons~=1.5.5
asgiref~=3.5.2
xlwt~=1.3.0
atlassian-python-api~=3.32.2

requests~=2.28.1